Rename handlers and split handlers/distgit.py
Rename DistGitPRHandler -> SourceGitPRToDistGitPRHandler
and move it to sourcegitPR_to_distgitPR.py

Rename
SyncFromGitlabMRHandler -> GitlabCIToSourceGitPRHandler
SyncFromPagurePRHandler -> PagureCIToSourceGitPRHandler
and move them to distgitCI_to_sourcegitPR.py

(This is really just a renaming/moving, no code change)
