+++ hardly/handlers/__init__.py
@@ -1,14 +1,15 @@
 # Copyright Contributors to the Packit project.
 # SPDX-License-Identifier: MIT
 
-from hardly.handlers.distgit import (
-    DistGitPRHandler,
-    SyncFromGitlabMRHandler,
-    SyncFromPagurePRHandler,
+from hardly.handlers.distgitCI_to_sourcegitPR import (
+    GitlabCIToSourceGitPRHandler,
+    PagureCIToSourceGitPRHandler,
 )
 
+from hardly.handlers.sourcegitPR_to_distgitPR import SourceGitPRToDistGitPRHandler
+
 __all__ = [
-    DistGitPRHandler.__name__,
-    SyncFromGitlabMRHandler.__name__,
-    SyncFromPagurePRHandler.__name__,
+    SourceGitPRToDistGitPRHandler.__name__,
+    GitlabCIToSourceGitPRHandler.__name__,
+    PagureCIToSourceGitPRHandler.__name__,
 ]
